revision-primer-proyecto

+++ mi_familia/models.py
@@ -3,5 +3,5 @@ from django.db import models
 class familia (models.Model):
     nombre=models.CharField(max_length=50)
     parentezco=models.CharField(max_length=50)
-    edad=models.FloatField()
+    edad=models.IntegerField()
     estudia=models.BooleanField()
